Update with alternative parser
The alternative parser essentially treats every line as a list element
as long as it does not have a recognizable key attached to it. Then it
generates the corresponding yaml format, and feeds it to pyyaml to do
the parsing. However the issue is that since pyyaml treats the lack of
'- ' as indication of a dictionary, the first level of indents will be
treated as dictionary and you can only have unique key values. But there
are duplicated key needed for first level and embedded indents.